Game in Python. Guess the right number.
Provided is the code in python for a basic number guesser game. The user provides the computer with a range of numbers and the computer randomly picks one of them. Now the user is supposed to guess which number has been picked. They win if they're able to guess it correctly and lose otherwise.